+++ Assets/_Core/Scenes/BallScene.unity
@@ -174,68 +174,15 @@ PrefabInstance:
       propertyPath: m_Name
       value: TopBorder
       objectReference: {fileID: 0}
+    - target: {fileID: 6675644798172047195, guid: d8cbed7de64a7a14c85ce8e7ed14fcd0, type: 3}
+      propertyPath: m_IsTrigger
+      value: 0
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: d8cbed7de64a7a14c85ce8e7ed14fcd0, type: 3}
---- !u!1001 &503650209
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    serializedVersion: 3
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4390203393362698855, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
-      propertyPath: m_Name
-      value: Ball
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-    m_RemovedGameObjects: []
-    m_AddedGameObjects: []
-    m_AddedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
 --- !u!1 &1542328745
 GameObject:
   m_ObjectHideFlags: 0
@@ -373,6 +320,71 @@ MonoBehaviour:
     m_MipBias: 0
     m_VarianceClampScale: 0.9
     m_ContrastAdaptiveSharpening: 0
+--- !u!1001 &1988322560
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1147724572960377830, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4390203393362698855, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: m_Name
+      value: Ball
+      objectReference: {fileID: 0}
+    - target: {fileID: 6687623290174318990, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: _speed
+      value: 6
+      objectReference: {fileID: 0}
+    - target: {fileID: 6687623290174318990, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
+      propertyPath: _initialSpeed
+      value: 40
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: b116bae02ee23db4c8468dc199b8f021, type: 3}
 --- !u!1001 &1988423147
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -561,4 +573,4 @@ SceneRoots:
   - {fileID: 1215186902720105925}
   - {fileID: 2077296271}
   - {fileID: 1988423147}
-  - {fileID: 503650209}
+  - {fileID: 1988322560}
